requirements.txt
litellm>=1.60.8
openai>=1.61.1
python-dotenv
anthropic>=0.45.0
requests
tqdm
pymongo>=4.0.0
redis>=4.0.0
upstash-vector
aiohttp
beautifulsoup4
